% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{i18n}
\alias{i18n}
\title{Create translator object}
\usage{
i18n(translations, language = NULL)
}
\arguments{
\item{translations}{Path to a json file containing translations}

\item{language}{The default language for the translation}
}
\description{
Create a new translator object
}
\examples{
path <- system.file("examples/simple.json", package = "traduire")
obj <- traduire::i18n(path)
obj$t("hello", language = "fr")
}
