% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{i18n}
\alias{i18n}
\title{Create translator object}
\usage{
i18n(resources, ..., options = NULL)
}
\arguments{
\item{resources}{Path to a json file containing translation
resources. If given in this way, then on-demand translation
loading (via \code{resource_pattern}) is disabled unless a
currently unexposed i18next option is used.}

\item{...}{Named options passed to \code{\link{traduire_options}}}

\item{options}{Options object passed to \code{\link{traduire_options}}}
}
\description{
Create a new translator object
}
\section{Warning}{


Note that the argument list here \emph{will} change.  The only
part of this that we consider stable is that the first argument
will represent a resource bundle.
}

\examples{
path <- system.file("examples/simple.json", package = "traduire")
obj <- traduire::i18n(path)
obj$t("hello", language = "fr")
}
