% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{lint_translations_report}
\alias{lint_translations_report}
\title{Show translation lint report}
\usage{
lint_translations_report(x, file = tempfile(fileext = ".html"))
}
\arguments{
\item{x}{A \code{lint_translations} object from
\code{\link{lint_translations}}}

\item{file}{File to save the report in (default is to use a
temporary file with a file extension of \code{.html}).}
}
\value{
This function is called for the side effect of viewing a
  report.  It will return (invisibly) the path to the produced
  report.
}
\description{
Report on translation linting.  At present this produces a
single-page html report that can be opened in a web browser.
Future versions will include reporting to the terminal.
}
